requirements.txt
tensorflow==1.2
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
mysql-connector-python
Werkzeug
Flask
Flask-HTTPAuth